refactor: Move CellState from view to models and update imports across controllers and views
